Gutenboarding: Extract to SignupDisplayNormal
