+++ client/landing/gutenboarding/components/signup-form/index.tsx
@@ -2,7 +2,7 @@
  * External dependencies
  */
 import React, { useState, useEffect } from 'react';
-import { Button, ExternalLink, TextControl, Modal, Notice } from '@wordpress/components';
+import { ExternalLink } from '@wordpress/components';
 import { useViewportMatch } from '@wordpress/compose';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { createInterpolateElement } from '@wordpress/element';
@@ -22,9 +22,7 @@ import {
 	useAnchorFmParams,
 	useIsAnchorFm,
 } from '../../path';
-import ModalSubmitButton from '../modal-submit-button';
 import './style.scss';
-import SignupFormHeader from './header';
 import {
 	initGoogleRecaptcha,
 	recordGoogleRecaptchaAction,
@@ -32,6 +30,7 @@ import {
 } from '../../lib/analytics';
 import { useTrackModal } from '../../hooks/use-track-modal';
 import config from '@automattic/calypso-config';
+import SignupDisplayNormal from './signup-display-normal';
 
 interface Props {
 	onRequestClose: () => void;
@@ -200,89 +199,20 @@ const SignupForm = ( { onRequestClose }: Props ) => {
 	const loginUrl = `/log-in/new${ langFragment }?redirect_to=${ loginRedirectUrl }&signup_url=${ signupUrl }`;
 
 	return (
-		<Modal
-			className={ 'signup-form' }
-			title={
-				isAnchorFmSignup
-					? __( 'Create your podcast site with WordPress.com' )
-					: __( 'Save your progress' )
-			}
-			onRequestClose={ closeModal }
-			focusOnMount={ false }
-			isDismissible={ false }
-			overlayClassName={ 'signup-form__overlay' }
-			// set to false so that 1password's autofill doesn't automatically close the modal
-			shouldCloseOnClickOutside={ false }
-		>
-			<SignupFormHeader onRequestClose={ closeModal } />
-
-			<div className="signup-form__body">
-				<h1 className="signup-form__title">
-					{ isAnchorFmSignup
-						? __( 'Create your podcast site with WordPress.com' )
-						: __( 'Save your progress' ) }
-				</h1>
-
-				<form onSubmit={ handleSignUp }>
-					<fieldset className="signup-form__fieldset">
-						<legend className="signup-form__legend">
-							<p>
-								{ isAnchorFmSignup
-									? __( 'Create a WordPress.com account and start creating your free site.' )
-									: __( 'Enter an email and password to save your progress and continue.' ) }
-							</p>
-						</legend>
-
-						<TextControl
-							value={ emailVal }
-							disabled={ isFetchingNewUser }
-							type="email"
-							onChange={ setEmailVal }
-							placeholder={ __( 'Email address' ) }
-							required
-							autoFocus={ ! isMobile } // eslint-disable-line jsx-a11y/no-autofocus
-						/>
-
-						<TextControl
-							value={ passwordVal }
-							disabled={ isFetchingNewUser }
-							type="password"
-							autoComplete="new-password"
-							onChange={ setPasswordVal }
-							placeholder={ __( 'Password' ) }
-							required
-						/>
-
-						{ errorMessage && (
-							<Notice className="signup-form__error-notice" status="error" isDismissible={ false }>
-								{ errorMessage }
-							</Notice>
-						) }
-
-						<div className="signup-form__footer">
-							<p className="signup-form__login-link">
-								<span>{ __( 'Already have an account?' ) }</span>{ ' ' }
-								<Button className="signup-form__link" isLink href={ loginUrl }>
-									{ __( 'Log in' ) }
-								</Button>
-							</p>
-
-							<p className="signup-form__link signup-form__terms-of-service-link">{ tos }</p>
-
-							<ModalSubmitButton disabled={ isFetchingNewUser } isBusy={ isFetchingNewUser }>
-								{ __( 'Create account' ) }
-							</ModalSubmitButton>
-
-							<p className="signup-form__link signup-form__terms-of-service-link signup-form__recaptcha_tos">
-								{ recaptcha_tos }
-							</p>
-						</div>
-					</fieldset>
-				</form>
-			</div>
-
-			<div id="g-recaptcha"></div>
-		</Modal>
+		<SignupDisplayNormal
+			closeModal={ closeModal }
+			emailVal={ emailVal }
+			errorMessage={ errorMessage }
+			handleSignUp={ handleSignUp }
+			isFetchingNewUser={ isFetchingNewUser }
+			isMobile={ isMobile }
+			loginUrl={ loginUrl }
+			passwordVal={ passwordVal }
+			recaptcha_tos={ recaptcha_tos }
+			setEmailVal={ setEmailVal }
+			setPasswordVal={ setPasswordVal }
+			tos={ tos }
+		/>
 	);
 };
